apps/backend/requirements-dev.txt
# Development Dependencies
-r requirements.txt

# Testing Framework
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==20.1.0
coverage==7.3.2
django-test-plus==2.3.0
model-bakery==1.17.0
freezegun==1.2.2
responses==0.24.1
httpretty==1.1.4
requests-mock==1.11.0

# Code Quality and Linting
black==23.11.0
isort==5.12.0
flake8==6.1.0
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-bugbear==23.12.2
flake8-comprehensions==3.15.0
flake8-simplify==0.20.0
mypy==1.7.1
django-stubs==4.2.7
djangorestframework-stubs==3.16.2
pre-commit==3.6.0
bandit==1.7.5
safety==2.3.5

# Development Tools
django-debug-toolbar==4.2.0
django-silk==5.0.4
ipython==8.17.2
jupyter==1.0.0

# Performance and Profiling
django-profiler==0.1.0
django-slowtests==0.1.0

# Migration Tools
django-test-migrations==1.3.0
django-migration-linter==3.0.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinx-autodoc-typehints==1.25.0
sphinxcontrib-django==2.4.0